requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==11.1.0
django-allauth==0.61.1
python-decouple==3.8
django-phonenumber-field[phonenumbers]==7.3.0
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-countries==7.5.1
django-filter==23.5
django-widget-tweaks==1.5.0
django-taggit==5.0.1
python-magic-bin==0.4.14
mysqlclient==2.2.7 